moved AssertLinks() into helpers in testutils package
