+++ authentication_handlers_test.go
@@ -89,7 +89,7 @@ func TestAuthenticationList(t *testing.T) {
 		}
 	}
 
-	AssertLinks(t, c.Request().RequestURI, out.Links, 100, 0)
+	testutils.AssertLinks(t, c.Request().RequestURI, out.Links, 100, 0)
 }
 
 func TestAuthenticationGet(t *testing.T) {
