+++ internal/testutils/helpers.go
@@ -1,12 +1,14 @@
 package testutils
 
 import (
+	"fmt"
 	"testing"
 
 	"github.com/RedHatInsights/sources-api-go/config"
 	"github.com/RedHatInsights/sources-api-go/internal/testutils/fixtures"
 	"github.com/RedHatInsights/sources-api-go/internal/testutils/parser"
 	"github.com/RedHatInsights/sources-api-go/model"
+	"github.com/RedHatInsights/sources-api-go/util"
 )
 
 var conf = config.Get()
@@ -52,3 +54,16 @@ func GetSourcesWithAppType(appTypeId int64) []model.Source {
 
 	return sources
 }
+
+func AssertLinks(t *testing.T, path string, links util.Links, limit int, offset int) {
+	expectedFirstLink := fmt.Sprintf("%s?limit=%d&offset=%d", path, limit, offset)
+	expectedLastLink := fmt.Sprintf("%s?limit=%d&offset=%d", path, limit, limit+offset)
+	if links.First != expectedFirstLink {
+		t.Error("first link is not correct for " + path)
+	}
+
+	if links.Last != expectedLastLink {
+		t.Error("last link is not correct for " + path)
+	}
+}
+
